Make sure Hypothetical actions allow Reveal
